+++ backend/ax_merchant_api/settings.py
@@ -67,6 +67,8 @@ INSTALLED_APPS = [
     "referrals",
     "chats",
     "whatsapp_messaging",
+    "subscriptions",
+    "devs",
 ]
 
 MIDDLEWARE = [