+++ backend/subscriptions/serializers.py
@@ -1,0 +1,22 @@
+from rest_framework import serializers
+from .models import SubscriptionInvoice, SubscriptionPlan, MerchantSubscription
+
+
+class SubscriptionPlanSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = SubscriptionPlan
+        fields = "__all__"
+
+
+class MerchantSubscriptionSerializer(serializers.ModelSerializer):
+    plan = SubscriptionPlanSerializer(read_only=True)
+
+    class Meta:
+        model = MerchantSubscription
+        fields = "__all__"
+
+
+class SubscriptionInvoiceSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = SubscriptionInvoice
+        fields = "__all__"